supp ticket v2

+++ Vue/vueAccueil.php
@@ -29,9 +29,10 @@
                                     </a><time><?= strftime("%a %d %b  %Y %H:%M:%S", strtotime($billet['TIC_DATE'])); ?></time>
                                     <!-- <i class="fa fa-comments"></i><a href="#">2 commentaires</a> -->
                                 </p>
-                                <a style="color:red"
+                                <a onclick="return confirm('&Ecirc;tes-vous sûr de vouloir supprimer ce ticket ?')"
+                                    style="color:red"
                                     href="index.php?action=supprimer&id=<?= $billet['id_billet'] ?>">Supprimer un
-                                    billet</a>
+                                    ticket</a>
                         </div>
                     </div>
                 </li>
